refactor(frontend): split flight details into focused subcomponents
