+++ apps/frontend/src/components/flights/details/FlightDetails.types.ts
@@ -1,0 +1,1 @@
+export type DownloadableFlightMedia = 'gpx' | 'video' | 'overlay';
